requirements.txt
# 全球情报系统 - Python 依赖
# Global Intelligence System - Python Dependencies
# 最后更新：2026-03-02

# ==================== 核心依赖 ====================
PyYAML>=6.0          # YAML 配置解析
croniter>=2.0.0      # Cron 表达式解析

# ==================== 数据采集 ====================
requests>=2.31.0     # HTTP 请求
feedparser>=6.0.10   # RSS 解析
beautifulsoup4>=4.12.0  # HTML 解析
lxml>=5.0.0          # XML/HTML 解析器

# ==================== 数据处理 ====================
# NLP 和实体识别
spacy>=3.7.0         # NLP 处理
# 安装模型：python -m spacy download zh_core_web_sm
# 安装模型：python -m spacy download en_core_web_sm

# 关系抽取和文本分析
transformers>=4.35.0 # Hugging Face 模型
torch>=2.1.0         # PyTorch

# 文本处理
nltk>=3.8.0          # 自然语言工具包
textblob>=0.17.1     # 情感分析

# ==================== 数据存储 ====================
# 关系数据库
sqlite3              # Python 内置（无需安装）
sqlalchemy>=2.0.0    # ORM 框架
alembic>=1.12.0      # 数据库迁移

# PostgreSQL（可选）
# psycopg2-binary>=2.9.0

# 图数据库
neo4j>=5.0.0         # Neo4j Python 驱动

# 缓存
redis>=5.0.0         # Redis 客户端

# ==================== 大模型接口 ====================
openai>=1.3.0        # OpenAI API 兼容接口
httpx>=0.25.0        # 异步 HTTP 客户端

# ==================== API 服务 ====================
fastapi>=0.104.0     # Web API 框架
uvicorn>=0.24.0      # ASGI 服务器
pydantic>=2.5.0      # 数据验证

# ==================== 任务队列（可选） ====================
# celery>=5.3.0      # 分布式任务队列
# rabbitmq           # 消息代理

# ==================== 监控和日志 ====================
# prometheus-client>=0.19.0  # 监控指标
# structlog>=23.2.0          # 结构化日志

# ==================== 测试 ====================
pytest>=7.4.0        # 测试框架
pytest-cov>=4.1.0    # 测试覆盖率
pytest-asyncio>=0.21.0  # 异步测试

# ==================== 开发工具 ====================
black>=23.0.0        # 代码格式化
flake8>=6.1.0        # 代码检查
mypy>=1.7.0          # 类型检查
pre-commit>=3.5.0    # Git 钩子

# ==================== 其他工具 ====================
python-dotenv>=1.0.0 # 环境变量管理
tenacity>=8.2.0      # 重试装饰器
aiohttp>=3.9.0       # 异步 HTTP

# ==================== 数据可视化（可选） ====================
# matplotlib>=3.8.0
# seaborn>=0.13.0
# plotly>=5.18.0
